requirements.txt
Pillow
Pydrive
wxpython